% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{bicycleClass}
\alias{bicycleClass}
\title{bicycleClass}
\usage{
bicycleClass(name, cyclepath, road, bicycleparking, area, editors, lasteditdate)
}
\arguments{
\item{name}{character}

\item{cyclepath}{length in km}

\item{road}{length in km}

\item{bicycleparking}{count of parking points}

\item{area}{hectares}

\item{editors}{total number editors of bicycle parking (optional)}

\item{lasteditdate}{Date of last edit of bicycleparking points (optional)}
}
\value{
dataframe
}
\description{
bicycleClass is used to calculate and classify an index for bicycle infrastructure based on a dataframe of
bicycle indicators.
}
\details{
For instance, the example dataset in this package uses data from Scotland and compares
it against Amsterdam. Unsuprisingly, Amsterdam provides the best infrastructre, the other Scottish
areas are classified based on 5 quintiles on a normalised scale between 0-1 (Amsterdam = 1).

This function is designed to work of value obtained from OpenStreetMap data. Specifically, it has been
designed to use the bicyclePostGis function. This function returns a dataframe from an OpenStreetMap
database with the arguments required for use with bicycleClass. Other data sources or
supplementary data sources could be used as long as the dataframe is constructed correctly for this function.
There are four indicator ratios of bicycle infrastructure which are combined to give an overall classification.
These are return in a dataframe: The ratio of cyclepath length to road length, ratio of National Cycle
Network routes to roads length and the amount of bicycle parking per hectare. The is also a
'ruralness weighting' applied to account for less densely populated areas which arguably have quieter roads
and require less cycle parking. The ruralness weighting is based on the relative amount of roads to area
as a proxy for population density. See bicyclePostGis function for detailed breakdown of the OpenStreetMap
data required for this function.

Once all indicators are calculated they are given weights 4:2:1 for Cyclepath, National Cycle Route
and Bicycle parking ratios respectively. This weigthting is based on expert opinion, bicycle literature
(subjective). Further work at linking bicycle indicators on outcomes i.e. the % of the public travelling
using bicycles is planned.

Each entry in the dataframe is also given a 'Confidence of Class'. In this context, it tries to measure
the sampling effort (number of OpenStreetMap editors) and the time of last edit to represent uncertainty in the
in the quality of the OpenStreetMap data.
}
\seealso{
bicyclePostGis
}

