% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bicycleData.R
\name{bicycleData}
\alias{bicycleData}
\title{Query bicycle data from a database which uses the osm2postgresql schema with all tags h-store activated.}
\usage{
bicycleStatus(user=user, password=password, dbname=dbname, host=host)
}
\arguments{
\item{user}{postgres user name}

\item{password}{postres password}

\item{dbname}{potsgres database name}

\item{host}{'localhost' or port}
}
\description{
requires OGR2OGR command line tool and access to osm2postgesql database schema
}

