% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bicycleStatus.R
\name{bicycleStatus}
\alias{bicycleStatus}
\title{bicycleStatus}
\format{A data frame with 13 variables in this column order:
\describe{
\item{\code{name}}{name or unique id of area}
\item{\code{cyclepath}}{length in km of cycle paths}
\item{\code{road}}{total length in km of paved roads}
\item{\code{bicycleparking}}{number of bicycle parking points}
\item{\code{area}}{Area in hectares of polygon being assessed}
\item{\code{routes}}{length in km of proposed Nation cycle route}
\item{\code{proposedroutes}}{length in km of proposed Nation cycle route}
\item{\code{proposedcyclepath}}{length in km of proposed cycle path}
\item{\code{constructioncyclepath}}{Length in km of cyclepath under construction}
\item{\code{editors}}{Number of openstreetmap editors of bicycle parking}
\item{\code{lasteditdate}}{Date of last edit in openstreetmap of bicycle parking}
\item{\code{proposedhighways}}{length in km of proposed roads}
\item{\code{constructionshighways}}{length in km of construction roads}
}}
\usage{
bicycleStatus(x)
}
\arguments{
\item{x}{Dataframe input from bicycleData function}

\item{amsterdamIndex}{Calculate status relative to data for Amsterdam included in
package (default=FALSE)}

\item{effort}{Estimate sampling effort to assess confidence in data quality
(default=TRUE)}
}
\value{
dataframe containing the following:
\describe{
\item{\code{name}}{name or unique id of area}
\item{\code{cyclepath to road
ratio}}{A ratio of cycle to road}
\item{\code{cyclepath to road ratio
norm}}{A ratio of cycle path to road normalise against max value or
Amsterdam=default}
\item{\code{Cycle path status}}{Status of cycle path based
on quintiles}
\item{\code{cyclepath to road ratio norm weighted}}{A ratio of
cycle path to road with * 4 weighting. The cycle path to raod ratio is deemed
to be *4 more important than the other indexes}
....to be completed
}
}
\description{
bicycleStatus is used to calculate and classify an index for bicycle
infrastructure based on a dataframe of bicycle indicators.
}
\details{
Using the example dataset in this package which uses data from Scotland and
compares it against Amsterdam, its possible to see that Amsterdam provides better
infrastructure. The Scottish areas are classified based on 5 quintiles
on a normalised scale between 0-1 (Amsterdam = 1).

This function is designed to work from indicator values obtained from
OpenStreetMap data. Specifically, it has been designed to use the bicycleData
function. This function returns a dataframe from an OpenStreetMap database
with the arguments required for use with bicycleStatus. Other data sources or
supplementary data sources could be used as long as the dataframe is
constructed correctly for this function. There are four indicator ratios of
bicycle infrastructure which are combined to give an overall classification.
These are return in a dataframe: The ratio of cyclepath length to road
length, ratio of National Cycle Network routes to roads length and the amount
of bicycle parking per hectare. The is also a 'ruralness weighting' applied
to account for less densely populated areas which arguably have quieter roads
and require less cycle parking. The ruralness weighting is based on the
relative amount of roads to area as a proxy for population density. See
bicycleData function for detailed breakdown of the OpenStreetMap data
required for this function.

Once all indicators are calculated they are given weights 4:2:1 for
Cyclepath, National Cycle Route and Bicycle parking ratios respectively. This
weigthting is based on expert opinion and bicycle literature (both of which
are deemed to subjective). Further work at linking bicycle indicators on
outcomes i.e. the % of the public travelling using bicycles is planned.
Ultimately, the index is base on quantitive data but the final assigned
status claculated for an area is based to some degree on the weighting given
to each indicator. The approach is to iterate the classification system to
include to closely align the status to socio-ecomonic-enviromental outcomes.

Each entry in the dataframe is also given a 'Confidence'. In this
context, it tries to measure the sampling effort (number of OpenStreetMap
editors) and the time of last edit to represent uncertainty in the
quality of the OpenStreetMap data. However, the current method is a fudge but
produce figures that appear roughly feasible. Further empirical testing
required to make this robust.

Objectives and measures are calculated to project the amount of improvement
required for a area to reach the equivalent status as the highest ranked area
within the dataset.
}
\examples{
bicycleStatus(scotlandAmsterdam)
}

