% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{bikr-package}
\alias{bikr-package}
\title{Generate bicycle outputs}
\description{
Based on data from OpenStreetmap data on a postgresql database using the
osmtoposgresql schema, this package classifies based on bicycle infrastruture
then then compare, constrast and project improvements requirements.
}
\details{
The only function you're likely to need from \pkg{bikr} is
\code{\link{bicycleStatus}}. Refer to the demo website to see how the package
works how to format the documentation.
}

