% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{max,biggestnumber}
\alias{test}
\title{max number}
\usage{
test(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
max value from input
}
\description{
Takes in numeric numbers and returns max
}

