% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{bikr-package}
\alias{bikr-package}
\title{Generate bicycle outputs}
\description{
Get query data on and Postgresql database using the osmtoposgresql schema and classify based
on bicycle infrastruture then then compare, constrast and project improvements requirements.
}
\details{
The only function you're likely to need from \pkg{bicycleStatus} is
\code{\link{bicycleStatus}}. To get the data from OpenStreetMap databaseuse \pkg{bicyleData} is
\code{\link{bicycleData}} Otherwise refer to the vignettes to see
how to format the documentation.
}

