% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikr-package.R
\docType{package}
\name{bikr-package}
\alias{bikr}
\alias{bikr-package}
\title{bikr: About the Package (One Line, Title Case)}
\usage{
library(bikr)
}
\description{
This package provides an example BikeShare trip dataset from Pittsburgh BikeShare (available for download at []()), as well as several scripts to analyze these trips. Package scripts can produce maps and statistical analysis (graphs, histograms) from trip data.

This package provides a range of functions varying in depth of analysis,
for a sample BikeShare Dataset available publicly
from (https://www.kaggle.com/adamscarroll/pittsburgh-healthy-ride-bikeshare-rentals-2021-q1/metadata).
The package also provides several tools and scripts for visualizing route distribution,
characteristics, lengths, as well as cross-analyzing all of these characteristics by ride-types.
}
\references{
\url{https://github.com/vanveghd/bikr}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/vanveghd/bikr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Daniel Van Veghel, School of Earth, Environment and Society, McMaster University \email{vanveghd@mcmaster.ca}
}
\keyword{Analysis,}
\keyword{Behaviour}
\keyword{BikeShare,}
\keyword{Travel}
\keyword{internal}
\keyword{statistics,}
