% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikr-package.R
\docType{package}
\name{bikr-package}
\alias{bikr}
\alias{bikr-package}
\title{bikr: An R package for simple Bike Share analysis}
\usage{
library(bikr)
}
\description{
This package provides an example BikeShare trip dataset from Pittsburgh BikeShare (available for download at []()), as well as several scripts to analyze these trips. Package scripts can produce maps and statistical analysis (graphs, histograms) from trip data.
}
\references{
\url{https://github.com/vanveghd/bikr}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{http://github.com/vanveghd/bikr/issues}
}

}
\author{
\strong{Maintainer}: Daniel Van Veghel \email{vanveghd@mcmaster.ca}


Daniel Van Veghel, School of Earth, Environment and Society, McMaster University \email{vanveghd@mcmaster.ca}
}
\keyword{Analysis,}
\keyword{Behaviour}
\keyword{BikeShare,}
\keyword{Travel}
\keyword{internal}
\keyword{statistics,}
